% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{daily_minimal_theme}
\alias{daily_minimal_theme}
\title{Daily Minimal Theme}
\usage{
daily_minimal_theme(
  base_size = 9,
  background = "#EEEEEE",
  textColor = "#101010",
  base_family = "Virtuous Slab"
)
}
\arguments{
\item{base_size}{base font size of ggplot theme}

\item{background}{the background color of plot}

\item{textColor}{the text color}

\item{base_family}{the textfont family}
}
\value{

}
\description{
Daily Minimal Theme
}
\examples{
library(ggplot2)
library(Kaashi)
square = regularPolygon(6)
tile <- motif(square,n = 6,theta = 45, delta = 0.3, polyLine = F, dist = 0.1)
p <- ggplot(tile)+
geom_sf(fill = textColor, color = textColor)+
  labs(caption = "Daily Motif \n \n NO. 1")+
  daily_minimal_theme()
}
