% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif.R
\name{motif}
\alias{motif}
\title{Design Tile Pattern}
\usage{
motif(
  box = rbind(c(-1, 0), c(1, 0), c(1, 2), c(-1, 2), c(-1, 0)),
  midpoint = c(0, 0),
  theta = 30,
  delta = 0.2,
  n = 4,
  dist = 0.001,
  circle = FALSE,
  radius = NULL,
  polyLine = F,
  drawBox = FALSE,
  cropBox = NULL
)
}
\arguments{
\item{box}{the polygon that the shape constructed by it.}

\item{midpoint}{is one point in boundary box that basis lines are constructed.}

\item{theta}{the angle of basis lines in degree between 0 to 90.}

\item{delta}{the distance of basis lines between 0 to 1 (it depend on boundary box).}

\item{n}{the number of sides of polygon.}

\item{dist}{is the size of interior lines.}

\item{circle}{is a Boolean variable that makes regions based on the intersection of circles instead of lines.}

\item{radius}{the double variable that if is not null, sets the line length or radius of the circle.}

\item{drawBox}{to show boundary box set it TRUE.}

\item{cropBox}{the polygon that the motif is cropped by it.}
}
\value{

}
\description{
Design Tile Pattern
}
\examples{
library(ggplot2)
tile <- motif(theta = 45, delta = 0.5, polyLine = T)
tilePlotter(tile)
tile <- motif(theta = 45, delta = 0.5, dist = 0.05, polyLine = F)
tilePlotter(tile)
s3 = sqrt(3)
hexagonal = rbind(c(-1,0), c(1,0), c(2,s3), c(1,2*s3), c(-1,2*s3),c(-2,s3),c(-1,0))
tile <- motif(theta = 45, n = 6, delta = 0.2, midpoint = c(0,0),
                box = hexagonal, drawBox = T, polyLine = T)
tilePlotter(tile)

}
