% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{motif_rotation}
\alias{motif_rotation}
\title{Motif Rotation}
\usage{
motif_rotation(tile, theta, center = c(0, 0))
}
\arguments{
\item{tile}{is sf object likes as output of motif function.}

\item{theta}{the angle of rotation.}

\item{center}{the center of rotation. it is 2-D vector like c(0,0).}
}
\value{
sf object
}
\description{
Motif Rotation
}
\examples{
library(ggplot2)
tile <- motif(polyLine = T, drawBox = T)
tile_30 <- motif_rotation(tile,30,center = c(0,1))
tilePlotter(tile)
tilePlotter(rbind(tile,tile_30))
}
