% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{motif_scale}
\alias{motif_scale}
\title{Motif Scale}
\usage{
motif_scale(tile, scale)
}
\arguments{
\item{tile}{is sf object likes as output of motif function.}

\item{scale}{numeric number that shape is scaled by it.}
}
\value{
sf object
}
\description{
Motif Scale
}
\examples{
library(ggplot2)
tile <- regularPolygon(3,sf = T)
scale_tile <- motif_scale(tile, 0.5)
tilePlotter(tile)
tilePlotter(rbind(tile, scale_tile))
}
