% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{motif_transfer}
\alias{motif_transfer}
\title{Motif Transfer}
\usage{
motif_transfer(tile, shift)
}
\arguments{
\item{tile}{is sf object likes as output of motif function.}

\item{shift}{is 2-D shift vector like c(1,1).}
}
\value{
sf object
}
\description{
Motif Transfer
}
\examples{
library(ggplot2)
tile <- motif(polyLine = T, drawBox = T)
tile_2_0 <- motif_transfer(tile, shift = c(2,0))
tilePlotter(tile)
tilePlotter(rbind(tile, tile_2_0))
}
