% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{motif_union}
\alias{motif_union}
\title{Union of two tile}
\usage{
motif_union(tile1, tile2)
}
\arguments{
\item{tile1}{sf object}

\item{tile2}{sf object}
}
\value{
sf object
}
\description{
Union of two tile
}
\examples{
library(ggplot2)
tile <- motif(polyLine = F, drawBox = T)
tile_2_0 <- motif_transfer(tile, shift = c(1,0))
tilePlotter(st_union(tile, tile_2_0),tileColor = c("red","yellow","blue"))
}
