% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tarh.R
\name{pattern}
\alias{pattern}
\title{pattern}
\usage{
pattern(
  theta = 30,
  delta = 0.2,
  n = 4,
  dist = 0.001,
  start_points = c(0, -1),
  polyLine = F,
  drawBox = FALSE,
  box = rbind(c(-1, -1), c(1, -1), c(1, 1), c(-1, 1), c(-1, -1))
)
}
\arguments{
\item{box}{}
}
\value{

}
\description{
pattern
}
