% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{regularPolygon}
\alias{regularPolygon}
\title{Create Regular Polygon}
\usage{
regularPolygon(n, sf = F)
}
\arguments{
\item{n}{number of sides}

\item{sf}{if TRUE the sf polygon object was returned.}
}
\value{
matrix of points or sf polygon.
}
\description{
Create Regular Polygon
}
\examples{
regularPolygon(5) \%>\% plot()
regularPolygon(5,TRUE) \%>\% tilePlotter()
}
