% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rotation}
\alias{rotation}
\title{Rotation function}
\usage{
rotation(theta)
}
\arguments{
\item{a}{angle of rotation in degree}
}
\value{
matrix of rotation
}
\description{
Rotation function
}
\examples{
rotation(45)
}
