% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{saveTile}
\alias{saveTile}
\title{Save Pattern}
\usage{
saveTile(plot, path, name, type = c("png", "eps", "svg", "pdf"), dpi = 200)
}
\arguments{
\item{plot}{name of ggplot object}

\item{path}{path of saving plot}

\item{name}{name of file c("png","eps","svg","pdf")}

\item{type}{includes subset of}

\item{dpi}{Plot resolution.}
}
\value{

}
\description{
Save Pattern
}
