% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tarh.R
\name{tarh}
\alias{tarh}
\title{Tarh}
\usage{
tarh(
  theta = 135,
  delta = 0.2,
  n = 4,
  start_points = c(0, -1),
  drawBox = FALSE,
  box = rbind(c(-1, -1), c(1, -1), c(1, 1), c(-1, 1), c(-1, -1))
)
}
\arguments{
\item{box}{}
}
\value{

}
\description{
Tarh
}
