% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kaashi.R
\name{tessellation}
\alias{tessellation}
\title{Kaashi}
\usage{
tessellation(
  tarh,
  n,
  d = 4,
  box = rbind(c(-1, -1), c(1, -1), c(1, 1), c(-1, 1), c(-1, -1))
)
}
\arguments{
\item{shift}{}
}
\value{

}
\description{
Kaashi
}
