% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{tilePlotter}
\alias{tilePlotter}
\title{Plot Tile and Tiling Pattern}
\usage{
tilePlotter(
  shape,
  tileColor = NULL,
  fill = "area",
  lineColor = "black",
  borderSize = 0.5,
  borderColor = "white"
)
}
\arguments{
\item{shape}{pattern or tessellation function outputs. It is sf class object.}

\item{tileColor}{if the shape is a polygon, for coloring its regions you can
set tileColor, which is a vector of colors equal to the number of regions.}

\item{fill}{is the name of shape sf column and default value is "area".
If you can color shape regions base of region code you can set it "name"}

\item{lineColor}{for coloring the shape of a class of polyline}

\item{borderSize}{the size of lines}

\item{borderColor}{the color of  shape's border of a class of polygon}
}
\value{
ggplot objects
}
\description{
plot is used for plot pattern or tessellation outputs. In fact it is ggplot-based
function that uses geom_sf to plot sf class object.
}
\examples{
library(ggplot2)
tile <- motif(theta = 45, delta = 0.5, polyLine = T)
tilePlotter(tile)
tiling <- tiling(tile, n = 5)
tilePlotter(tiling)
tile <- motif(theta = 45, delta = 0.5, dist = 0.05, polyLine = F)
tiling <- tiling(tile, n = 5)
tilePlotter(tiling,tileColor = c("#EF821D","#2eb6c2","#0C3263"),borderSize = 0.01)
}
