% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pattern.R
\name{pattern}
\alias{pattern}
\title{Design Tile Pattern}
\usage{
pattern(
  theta = 30,
  delta = 0.2,
  n = 4,
  dist = 0.001,
  start_points = c(0, -1),
  polyLine = F,
  drawBox = FALSE,
  box = rbind(c(-1, -1), c(1, -1), c(1, 1), c(-1, 1), c(-1, -1))
)
}
\arguments{
\item{theta}{the angle of basis lines in degree between 0 to 90.}

\item{delta}{the distance of basis lines between 0 to 1 (it depend on boundary box).}

\item{n}{the number of sides of polygon.}

\item{dist}{is the size of interior lines.}

\item{start_points}{is one point in boundary box that basis lines are constructed.}

\item{drawBox}{to show boundary box set it TRUE.}

\item{box}{the polygon that the shape constructed by it.}
}
\value{

}
\description{
Design Tile Pattern
}
\examples{
pattern(theta = 45, delta = 0.5, polyLine = T)
pattern(theta = 45, delta = 0.5, dist = 0.05, polyLine = F)
s3 = sqrt(3)
hexagonal = rbind(c(-1,0), c(1,0), c(2,s3), c(1,2*s3), c(-1,2*s3),c(-2,s3),c(-1,0))
pattern(theta = 45, n = 6, delta = 0.2, start_points = c(0,0),
                box = hexagonal, drawBox = T, polyLine = T)

}
